DRY RegistrationTest setup/teardown to a module.
