+++ test/controllers/v1/registrations_controller_test.rb
@@ -1,17 +1,7 @@
 require 'test_helper'
 
 class V1::RegistrationsControllerTest < ActionController::TestCase
-  def setup
-    @event = Event.create(name: "Moe's bar", date: Time.zone.now)
-    @invitation = Invitation.create(event: @event, email: 'johndoe@example.com')
-    @user_data = { email: @invitation.email, password: 'my-password' }
-  end
-
-  def teardown
-    Event.delete_all
-    Invitation.delete_all
-    User.delete_all
-  end
+  include RegistrationSetup
 
   test 'create from invitation success' do
     assert_difference 'User.count' do
